refactor: rename type name and use getVideo on /category/video page
